requirements.txt
pymongo
gradio
langchain_ollama 
langgraph 
langchain_together 
trustcall 
gradio 
langchain_community 
pypdf
huggingface_hub==0.25.2
pydantic==2.10.6
python-dotenv
langgraph-store-mongodb